requirements.txt
alembic
asyncpg
factory-boy
faker
fastapi
httpx
jinja2
langcodes
pydantic
pydantic-settings
pycryptodome
PyJWT
pytest
pytest-asyncio
redis
rfc3986
sqlalchemy
uvicorn
